Changed admissibility parameter to floating point (double)

+++ include/hicma/classes/hierarchical.h
@@ -64,7 +64,7 @@ class Hierarchical : public Matrix {
     int64_t n_rows, int64_t n_cols,
     int64_t rank,
     int64_t nleaf,
-    int64_t admis=1,
+    double admis=1,
     int64_t n_row_blocks=2, int64_t n_col_blocks=2,
     int admis_type=POSITION_BASED_ADMIS,
     int64_t row_start=0, int64_t col_start=0
@@ -74,7 +74,7 @@ class Hierarchical : public Matrix {
     Dense&& A,
     int64_t rank,
     int64_t nleaf,
-    int64_t admis=1,
+    double admis=1,
     int64_t n_row_blocks=2, int64_t n_col_blocks=2,
     int admis_type=POSITION_BASED_ADMIS,
     int64_t row_start=0, int64_t col_start=0
@@ -86,7 +86,7 @@ class Hierarchical : public Matrix {
     int64_t n_rows, int64_t n_cols,
     int64_t rank,
     int64_t nleaf,
-    int64_t admis=1,
+    double admis=1,
     int64_t n_row_blocks=2, int64_t n_col_blocks=2,
     int basis_type=NORMAL_BASIS, int admis_type=POSITION_BASED_ADMIS,
     int64_t row_start=0, int64_t col_start=0
